+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.13.0] - 2024-04-26
+
+- Add support for recursive liquid tags [#1731](https://github.com/Shopify/liquid/pull/1731)
+- Match rounding precision to ruby gem when using the `round` filter with invalid precision.
+
 ## [0.12.0] - 2024-01-24
 
 - Add support for inline comments